Convert pay frequency from weekly to semi-monthly
Why? Because it makes monthly budgeting and cash flow analysis more predictable. The problem with a weekly frequency is that some months there are 4 paydays and some months there are 5. Because most budgeting happens on a monthly cycle, this leads to a lot of unnecessary volatility from month-to-month—for both the company and its employees.